+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Prompt when Omeka version does not fit with plugin version.
+- Repository name during plugin selection.
 
 ## [0.5.2] - 2017-07-03
 ### Added